requirements.txt
streamlit==1.36.0
streamlit-extras==0.4.0
google-generativeai==0.7.2
langchain==0.2.12
langchain-core==0.2.28
langchain-google-genai==1.0.7
langgraph==0.2.14
pydantic==2.7.4
python-dotenv==1.0.1
faiss-cpu==1.8.0
tenacity==8.3.0
pygments==2.18.0
